i2c.c: add i2c read/write functions
i2c_writebytes() needs review; I've used a little trick here: the
first element of the bytes array is passed as the command byte.

i2c_readbyte() is non-essential; it is typically used for getting an
ACK from a particular slave, to check if it is present/awake on the bus.