i2c.c: add functionality to support combined read+write
i2c_add_to_buf and i2c_readwrite solve most of my problems (documented
in a blog entry).

i2c_add_to_buf() takes the address of the slave, the operation (read/write),
a pointer to the buffer and the length as *arguments* and adds it to a
global buffer of struct i2c_msg. Currently, the maximum size of the buffer
is 2, limiting the number of concurrent reads or writes to 2. However,
it doesn't limit the size of the individual read or write buffer.

i2c_readwrite() simply calls an ioctl command on I2C_RDWR and returns the
status of the operation.